requirements.txt
Flask
Flask-PyMongo
flask-cors
pymongo
dnspython
python-dotenv
gunicorn
